Wien2k parser: fix eingenvalues parsing
The problem is that there can be different amount of eigenvalues
calculated at every k-point. However, the metainfo expects a homogeneous
array, so just find the k-point with the least amount and discard
the extra values at other k-points.

The current test worked just due to sheer luck. There were 116
eigenvalues at some k-points and 114 at others, but since it resulted
in an exact average of 115, the test was happy. However some eigenvalues
were incorrectly atributed to incorrect k-point.

Also the sorting was somehow broken, just sort the files before reading
and the k-points are than guaranteed to come in order. Its way better to
sort few strings instead of numpy array with pottentially thousands of
values anyway.